pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*"]

[project]
name = "resume-ats-optimizer"
version = "0.1.0"
description = "Resume ATS Optimizer with AG2 multi-agent pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "anthropic>=0.42.0",
    "openai>=1.60.0",
    "chromadb>=0.6.0",
    "ag2[openai]>=0.6.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
